+++ packages/core/src/i18n/Translations.ts
@@ -17,9 +17,12 @@ limitations under the License.
 */
 
 import Client from '../Client';
-import { NONE } from './Constants';
-import { get, load } from './MaxXmlRequest';
-import type MaxXmlRequest from './MaxXmlRequest';
+import { NONE } from '../util/Constants';
+import { get, load } from '../util/MaxXmlRequest';
+import type MaxXmlRequest from '../util/MaxXmlRequest';
+import { TranslationsConfig } from './config';
+
+// mxGraph source code: https://github.com/jgraph/mxgraph/blob/v4.2.2/javascript/src/js/util/mxResources.js
 
 /**
  * Implements internationalization. You can provide any number of
@@ -83,32 +86,33 @@ class Translations {
   static extension = '.txt';
 
   /**
-   * Specifies whether values in resource files are encoded with \u or
-   * percentage. Default is false.
+   * Specifies whether values in resource files are encoded with `\u` or percentage.
+   * @default false
    */
   static resourcesEncoded = false;
 
   /**
    * Specifies if the default file for a given basename should be loaded.
-   * Default is true.
+   * @default true
    */
   static loadDefaultBundle = true;
 
   /**
-   * Specifies if the specific language file file for a given basename should
-   * be loaded. Default is true.
+   * Specifies if the specific language file for a given basename should be loaded.
+   * @default true
    */
   static loadSpecialBundle = true;
 
   /**
-   * Hook for subclassers to disable support for a given language. This
-   * implementation returns true if lan is in <Client.languages>.
+   * Hook for subclassers to disable support for a given language.
+   * This implementation returns `true` if `lan` is in {@link TranslationsConfig.languages}.
    *
    * @param lan The current language.
    */
   static isLanguageSupported = (lan: string): boolean => {
-    if (Client.languages != null) {
-      return Client.languages.indexOf(lan) >= 0;
+    const languages = TranslationsConfig.getLanguages();
+    if (languages) {
+      return languages.indexOf(lan) >= 0;
     }
     return true;
   };
@@ -131,21 +135,18 @@ class Translations {
   /**
    * Hook for subclassers to return the URL for the special bundle. This
    * implementation returns `basename + '_' + lan + <extension>` or `null` if
-   * {@link Translations.loadSpecialBundle} is `false` or `lan` equals {@link Client.defaultLanguage}.
+   * {@link loadSpecialBundle} is `false` or `lan` equals {@link TranslationsConfig.getDefaultLanguage}.
    *
-   * If {@link Translations#languages} is not null and {@link Client.language} contains
-   * a dash, then this method checks if {@link Translations.isLanguageSupported} returns `true`
+   * If {@link TranslationsConfig.getLanguages} is not `null` and {@link TranslationsConfig.getLanguage} contains
+   * a dash, then this method checks if {@link isLanguageSupported} returns `true`
    * for the full language (including the dash). If that returns false the
    * first part of the language (up to the dash) will be tried as an extension.
-   *
-   * If {@link Translations#language} is null then the first part of the language is
-   * used to maintain backwards compatibility.
    *
    * @param basename The basename for which the file should be loaded.
    * @param lan The language for which the file should be loaded.
    */
   static getSpecialBundle = (basename: string, lan: string): string | null => {
-    if (Client.languages == null || !Translations.isLanguageSupported(lan)) {
+    if (!TranslationsConfig.getLanguages() || !Translations.isLanguageSupported(lan)) {
       const dash = lan.indexOf('-');
 
       if (dash > 0) {
@@ -156,7 +157,7 @@ class Translations {
     if (
       Translations.loadSpecialBundle &&
       Translations.isLanguageSupported(lan) &&
-      lan != Client.defaultLanguage
+      lan != TranslationsConfig.getDefaultLanguage()
     ) {
       return `${basename}_${lan}${Translations.extension}`;
     }
@@ -186,8 +187,7 @@ class Translations {
     lan: string | null = null,
     callback: Function | null = null
   ): void => {
-    lan =
-      lan != null ? lan : Client.language != null ? Client.language.toLowerCase() : NONE;
+    lan ??= TranslationsConfig.getLanguage()?.toLowerCase() ?? NONE;
 
     if (lan !== NONE) {
       const defaultBundle = Translations.getDefaultBundle(basename, lan);